% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_catalog.R
\name{createCatalog}
\alias{createCatalog}
\title{Create a catalog}
\usage{
createCatalog(verbose = TRUE)
}
\value{
nothing. The catalog is stored as an in-package  object.
}
\description{
Check whether a catalog file exists, if not create and write to repository a new catalog.
}
