% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_catalog.R
\name{getCatalog}
\alias{getCatalog}
\title{Access catalog data}
\usage{
getCatalog()
}
\value{
a data frame.
}
\description{
exports the repository catalog as a dataframe.

Note that the catalog is stored as a package option, i.e. as an inaccessible in-package object, this functions males a copy of the catalog.
}
