% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{getEXIFData}
\alias{getEXIFData}
\title{extract EXIF data}
\usage{
getEXIFData(EXIFDir = getwd(), tz = Sys.timezone(), offset = 0,
  verbose = FALSE)
}
\arguments{
\item{EXIFDir}{character, a \code{cameratraps} "sd card directory", i.e. a directory where camera trap files are stored.}

\item{tz}{character, see \link{OlsonNames}: a valit time zone designator, will be used ot fix EXIF timestamps accordin ti the time zone where the camera trap operated.}

\item{offset}{numeric, a time offset in hours (decimal hours) that will be algebtically added to EXIF timestamps. Use with care.}
}
\value{
a dataframe containing file names (\code{Raw.Names}), a timestamp (\code{Photo.Time}), the type of file (\code{Photo.Type}, and the "camera directory" where each file is stored (\code{Sampling.Event}).
}
\description{
get EXIF data from image and video files in a given directory
}
