% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{getRepository}
\alias{getRepository}
\title{Get the repository path}
\usage{
getRepository()
}
\value{
A character string or \code{NULL} and an error message, if a repository path has not been set.
}
\description{
Return the current \code{cameratraps} repository root path
}
\seealso{
\link{setRepository}
}
