% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_catalog.R
\name{listCameraDir}
\alias{listCameraDir}
\title{list repository structures}
\usage{
listCameraDir(siteDirName)
}
\description{
list camera directories in a repository
}
\seealso{
Other repository functions: \code{\link{listDataDir}},
  \code{\link{listSiteDir}}
}
\concept{repository functions}
