% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_catalog.R
\name{listDataDir}
\alias{listDataDir}
\title{list repository structures}
\usage{
listDataDir(siteDirName, cameraDirName)
}
\description{
list sd card data directories in a repository
}
\seealso{
Other repository functions: \code{\link{listCameraDir}},
  \code{\link{listSiteDir}}
}
\concept{repository functions}
